requirements.txt
alembic==1.0.11
amqp==2.5.1
anyapi==1.1.401
asn1crypto==0.24.0
autopep8==1.4.4
Babel==2.7.0
beautifulsoup4==4.8.0
billiard==3.6.1.0
blinker==1.4
bs4==0.0.1
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.7
cycler==0.10.0
emoji==0.5.3
Flask==1.1.1
Flask-Admin==1.5.3
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
flask-paginate==0.5.3
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
future==0.17.1
idna==2.8
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
kombu==4.6.4
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mp3-tagger==1.0
mysqlclient==1.4.4
numpy==1.17.0
passlib==1.7.1
pep8==1.7.1
pkg-resources==0.0.0
praw==6.3.1
prawcore==1.0.1
pycodestyle==2.5.0
pycparser==2.19
pyparsing==2.4.2
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-telegram-bot==12.0.0b1
pytz==2019.2
redis==3.3.7
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
speaklater==1.3
SQLAlchemy==1.3.7
timeago==1.0.10
tornado==6.0.3
update-checker==0.16
urllib3==1.25.3
vine==1.3.0
websocket-client==0.56.0
Werkzeug==0.15.5
WTForms==2.2.1
youtube-dl==2019.8.13
zipp==0.5.2
